Move middleware into lib folder
This moves the `middleware` folder into the `lib` folder and updates file references to them.

This is to enable easier updates to the kit in future, as the main folder to copy across will just be the `lib` one.
