+++ lib/middleware/authentication.js
@@ -1,5 +1,5 @@
 const url = require('url');
-const { encryptPassword } = require('../lib/utils');
+const { encryptPassword } = require('../utils');
 
 const allowedPathsWhenUnauthenticated = [
   '/prototype-admin/password',
