+++ lib/middleware/auto-routing.js
