+++ lib/middleware/prototype-admin-routes.js
@@ -4,7 +4,7 @@ const router = express.Router();
 
 const password = process.env.PROTOTYPE_PASSWORD;
 
-const { encryptPassword } = require('../lib/utils');
+const { encryptPassword } = require('../utils');
 
 router.get('/password', (req, res) => {
   const returnURL = req.query.returnURL || '/';
